python_scripts/requirements.txt
openai-whisper
yt-dlp
spacy
nltk
